[Fix]: Fix GetGoals

+++ core/database/src/main/java/com/hmh/hamyeonham/core/database/model/UsageGoalsEntity.kt
@@ -1,11 +1,10 @@
 package com.hmh.hamyeonham.core.database.model
 
-import androidx.room.Dao
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "usage_goals")
 data class UsageGoalsEntity(
     @PrimaryKey val packageName: String,
-    val goalTime: Int
+    val goalTime: Long,
 )
